make the user agent name getter a generic utility
